refactor: split session tests by persistence, queue IPC, and turn control
