+++ test/session-persistence.test.ts
@@ -1,17 +1,10 @@
 import assert from "node:assert/strict";
 import { spawn } from "node:child_process";
-import { createHash } from "node:crypto";
 import { once } from "node:events";
 import fs from "node:fs/promises";
-import net from "node:net";
 import os from "node:os";
 import path from "node:path";
 import test from "node:test";
-import type { SetSessionConfigOptionResponse } from "@agentclientprotocol/sdk";
-import {
-  QueueOwnerTurnController,
-  type QueueOwnerActiveSessionController,
-} from "../src/queue-owner-turn-controller.js";
 import type { SessionRecord } from "../src/types.js";
 
 type SessionModule = typeof import("../src/session.js");
@@ -432,305 +425,6 @@ test("normalizeQueueOwnerTtlMs applies default and edge-case normalization", asy
     assert.equal(session.normalizeQueueOwnerTtlMs(15_000), 15_000);
   });
 });
-
-test("cancelSessionPrompt sends cancel request to active queue owner", async () => {
-  await withTempHome(async (homeDir) => {
-    const session = await loadSessionModule();
-    const sessionId = "cancel-session";
-    const keeper = spawn(process.execPath, ["-e", "setInterval(() => {}, 1000);"], {
-      stdio: "ignore",
-    });
-    await once(keeper, "spawn");
-    const queueDir = path.join(homeDir, ".acpx", "queues");
-    await fs.mkdir(queueDir, { recursive: true });
-
-    const queueKey = createHash("sha256").update(sessionId).digest("hex").slice(0, 24);
-    const socketPath =
-      process.platform === "win32"
-        ? `\\\\.\\pipe\\acpx-${queueKey}`
-        : path.join(queueDir, `${queueKey}.sock`);
-    const lockPath = path.join(queueDir, `${queueKey}.lock`);
-
-    await fs.writeFile(
-      lockPath,
-      `${JSON.stringify({
-        pid: keeper.pid,
-        sessionId,
-        socketPath,
-      })}\n`,
-      "utf8",
-    );
-
-    const server = net.createServer((socket) => {
-      socket.setEncoding("utf8");
-      let buffer = "";
-      socket.on("data", (chunk: string) => {
-        buffer += chunk;
-        const newlineIndex = buffer.indexOf("\n");
-        if (newlineIndex < 0) {
-          return;
-        }
-        const line = buffer.slice(0, newlineIndex).trim();
-        if (!line) {
-          return;
-        }
-
-        const request = JSON.parse(line) as { requestId: string; type: string };
-        assert.equal(request.type, "cancel_prompt");
-        socket.write(
-          `${JSON.stringify({
-            type: "accepted",
-            requestId: request.requestId,
-          })}\n`,
-        );
-        socket.write(
-          `${JSON.stringify({
-            type: "cancel_result",
-            requestId: request.requestId,
-            cancelled: true,
-          })}\n`,
-        );
-        socket.end();
-      });
-    });
-
-    await new Promise<void>((resolve, reject) => {
-      const onError = (error: Error) => {
-        reject(error);
-      };
-      server.once("error", onError);
-      server.listen(socketPath, () => {
-        server.off("error", onError);
-        resolve();
-      });
-    });
-
-    try {
-      const result = await session.cancelSessionPrompt({ sessionId });
-      assert.equal(result.cancelled, true);
-      assert.equal(result.sessionId, sessionId);
-    } finally {
-      await new Promise<void>((resolve) => {
-        server.close(() => resolve());
-      });
-      if (process.platform !== "win32") {
-        await fs.rm(socketPath, { force: true });
-      }
-      if (keeper.pid && keeper.exitCode == null && keeper.signalCode == null) {
-        keeper.kill("SIGKILL");
-      }
-    }
-  });
-});
-
-test("QueueOwnerTurnController tracks explicit lifecycle states", async () => {
-  const controller = createQueueOwnerTurnController();
-  assert.equal(controller.lifecycleState, "idle");
-
-  controller.beginTurn();
-  assert.equal(controller.lifecycleState, "starting");
-
-  controller.markPromptActive();
-  assert.equal(controller.lifecycleState, "active");
-
-  controller.endTurn();
-  assert.equal(controller.lifecycleState, "idle");
-
-  controller.beginClosing();
-  assert.equal(controller.lifecycleState, "closing");
-  const cancelled = await controller.requestCancel();
-  assert.equal(cancelled, false);
-});
-
-test("QueueOwnerTurnController cancels immediately for active prompts", async () => {
-  const controller = createQueueOwnerTurnController();
-  let cancelCalls = 0;
-
-  controller.beginTurn();
-  controller.setActiveController(
-    makeActiveController({
-      hasActivePrompt: () => true,
-      requestCancelActivePrompt: async () => {
-        cancelCalls += 1;
-        return true;
-      },
-    }),
-  );
-  controller.markPromptActive();
-
-  const cancelled = await controller.requestCancel();
-  assert.equal(cancelled, true);
-  assert.equal(cancelCalls, 1);
-  assert.equal(controller.hasPendingCancel, false);
-});
-
-test("QueueOwnerTurnController defers cancel while turn is starting", async () => {
-  const controller = createQueueOwnerTurnController();
-  let promptActive = false;
-  let cancelCalls = 0;
-
-  controller.beginTurn();
-  controller.setActiveController(
-    makeActiveController({
-      hasActivePrompt: () => promptActive,
-      requestCancelActivePrompt: async () => {
-        cancelCalls += 1;
-        return promptActive;
-      },
-    }),
-  );
-
-  const accepted = await controller.requestCancel();
-  assert.equal(accepted, true);
-  assert.equal(cancelCalls, 0);
-  assert.equal(controller.hasPendingCancel, true);
-
-  const beforeActive = await controller.applyPendingCancel();
-  assert.equal(beforeActive, false);
-  assert.equal(cancelCalls, 0);
-  assert.equal(controller.hasPendingCancel, true);
-
-  promptActive = true;
-  controller.markPromptActive();
-  const afterActive = await controller.applyPendingCancel();
-  assert.equal(afterActive, true);
-  assert.equal(cancelCalls, 1);
-  assert.equal(controller.hasPendingCancel, false);
-});
-
-test("QueueOwnerTurnController routes setSessionMode through active controller", async () => {
-  let activeCalls = 0;
-  let fallbackCalls = 0;
-  const observedTimeouts: Array<number | undefined> = [];
-  const fallbackTimeouts: Array<number | undefined> = [];
-
-  const controller = createQueueOwnerTurnController({
-    withTimeout: async (run, timeoutMs) => {
-      observedTimeouts.push(timeoutMs);
-      return await run();
-    },
-    setSessionModeFallback: async (_modeId, timeoutMs) => {
-      fallbackCalls += 1;
-      fallbackTimeouts.push(timeoutMs);
-    },
-  });
-
-  controller.setActiveController(
-    makeActiveController({
-      setSessionMode: async () => {
-        activeCalls += 1;
-      },
-    }),
-  );
-
-  await controller.setSessionMode("plan", 1250);
-  assert.equal(activeCalls, 1);
-  assert.equal(fallbackCalls, 0);
-  assert.deepEqual(observedTimeouts, [1250]);
-  assert.deepEqual(fallbackTimeouts, []);
-});
-
-test("QueueOwnerTurnController routes setSessionConfigOption through fallback when inactive", async () => {
-  let fallbackCalls = 0;
-  const fallbackTimeouts: Array<number | undefined> = [];
-
-  const controller = createQueueOwnerTurnController({
-    setSessionConfigOptionFallback: async (_configId, _value, timeoutMs) => {
-      fallbackCalls += 1;
-      fallbackTimeouts.push(timeoutMs);
-      return { configOptions: [] };
-    },
-  });
-
-  const response = await controller.setSessionConfigOption(
-    "approval_policy",
-    "strict",
-    2300,
-  );
-  assert.equal(fallbackCalls, 1);
-  assert.deepEqual(fallbackTimeouts, [2300]);
-  assert.deepEqual(response, { configOptions: [] });
-});
-
-test("QueueOwnerTurnController rejects control requests while closing", async () => {
-  let setModeFallbackCalls = 0;
-  let setConfigFallbackCalls = 0;
-  const controller = createQueueOwnerTurnController({
-    setSessionModeFallback: async () => {
-      setModeFallbackCalls += 1;
-    },
-    setSessionConfigOptionFallback: async () => {
-      setConfigFallbackCalls += 1;
-      return { configOptions: [] };
-    },
-  });
-
-  controller.beginClosing();
-
-  await assert.rejects(
-    async () => await controller.setSessionMode("plan"),
-    /Queue owner is closing/,
-  );
-  await assert.rejects(
-    async () => await controller.setSessionConfigOption("k", "v"),
-    /Queue owner is closing/,
-  );
-  assert.equal(setModeFallbackCalls, 0);
-  assert.equal(setConfigFallbackCalls, 0);
-});
-
-type QueueOwnerTurnControllerOverrides = Partial<{
-  withTimeout: <T>(run: () => Promise<T>, timeoutMs?: number) => Promise<T>;
-  setSessionModeFallback: (modeId: string, timeoutMs?: number) => Promise<void>;
-  setSessionConfigOptionFallback: (
-    configId: string,
-    value: string,
-    timeoutMs?: number,
-  ) => Promise<SetSessionConfigOptionResponse>;
-}>;
-
-function createQueueOwnerTurnController(
-  overrides: QueueOwnerTurnControllerOverrides = {},
-): QueueOwnerTurnController {
-  const withTimeout =
-    overrides.withTimeout ??
-    (async <T>(run: () => Promise<T>): Promise<T> => await run());
-  const setSessionModeFallback =
-    overrides.setSessionModeFallback ??
-    (async (): Promise<void> => {
-      // no-op
-    });
-  const setSessionConfigOptionFallback =
-    overrides.setSessionConfigOptionFallback ??
-    (async () => ({
-      configOptions: [],
-    }));
-
-  return new QueueOwnerTurnController({
-    withTimeout,
-    setSessionModeFallback,
-    setSessionConfigOptionFallback,
-  });
-}
-
-type ActiveControllerOverrides = Partial<QueueOwnerActiveSessionController>;
-
-function makeActiveController(
-  overrides: ActiveControllerOverrides = {},
-): QueueOwnerActiveSessionController {
-  return {
-    hasActivePrompt: overrides.hasActivePrompt ?? (() => false),
-    requestCancelActivePrompt:
-      overrides.requestCancelActivePrompt ?? (async () => false),
-    setSessionMode:
-      overrides.setSessionMode ??
-      (async () => {
-        // no-op
-      }),
-    setSessionConfigOption:
-      overrides.setSessionConfigOption ?? (async () => ({ configOptions: [] })),
-  };
-}
 
 async function loadSessionModule(): Promise<SessionModule> {
   const cacheBuster = `${Date.now()}-${Math.random()}`;
